+++ pybraries/__init__.py
@@ -1,3 +1,4 @@
+"""Module that imports helpers"""
 from .helpers import sess
 from .make_request import make_request
 from .pagination import fix_pages
@@ -5,3 +6,5 @@ from .search import Search
 from .search_helpers import search_api
 from .subscribe import Subscribe
 from .subscription_helpers import sub_api
+
+__all__ = ["sess", "make_request", "fix_pages", "Search", "search_api", "Subscribe", "sub_api"]