+++ pybraries/search.py
@@ -4,14 +4,12 @@ from typing import Any
 from pybraries.search_helpers import search_api
 
 
-class Search(object):
+class Search:
     """Class for wrapping the libraries.io API for
     platform, project, repo, and user GET actions"""
 
-    def __init__(self):
-        pass
-
-    def platforms(self) -> Any:
+    @staticmethod
+    def platforms() -> Any:
         """
         Return a list of supported package managers.
 
@@ -21,7 +19,8 @@ class Search(object):
 
         return search_api("platforms")
 
-    def project(self, platforms: str, name: str) -> Any:
+    @staticmethod
+    def project(platforms: str, name: str) -> Any:
         """
         Return information about a project and its versions from a platform (e.g. PyPI).
 
@@ -31,17 +30,14 @@ class Search(object):
         Returns:
             List of dictionaries with information about the project from libraries.io.
         """
-
-        null = None  # null comes in looking like a variable for some platforms
         return search_api("project", platforms, name)
 
-    def project_dependencies(
-        self, platforms: str, project: str, version: str = None
-    ) -> Any:
+    @staticmethod
+    def project_dependencies(platforms: str, project: str, version: str = None) -> Any:
         """
         Get dependencies for a version of a project.
 
-        Returns latest version info.
+        Returns the latest version info.
 
         Args:
             platforms: package manager (e.g. "pypi").
@@ -53,9 +49,8 @@ class Search(object):
 
         return search_api("project_dependencies", platforms, project, version=version)
 
-    def project_dependents(
-        self, platforms: str, project: str, version: str = None
-    ) -> Any:
+    @staticmethod
+    def project_dependents(platforms: str, project: str, version: str = None) -> Any:
         """
         Get projects that have at least one version that depends on a given project.
 
@@ -69,7 +64,8 @@ class Search(object):
 
         return search_api("project_dependents", platforms, project, version=version)
 
-    def project_dependent_repositories(self, platforms: str, project: str) -> Any:
+    @staticmethod
+    def project_dependent_repositories(platforms: str, project: str) -> Any:
         """
         Get repositories that depend on a given project.
 
@@ -82,7 +78,8 @@ class Search(object):
 
         return search_api("project_dependent_repositories", platforms, project)
 
-    def project_contributors(self, platforms: str, project: str) -> Any:
+    @staticmethod
+    def project_contributors(platforms: str, project: str) -> Any:
         """
         Get users that have contributed to a given project.
 
@@ -95,7 +92,8 @@ class Search(object):
 
         return search_api("project_contributors", platforms, project)
 
-    def project_sourcerank(self, platforms: str, project: str) -> Any:
+    @staticmethod
+    def project_sourcerank(platforms: str, project: str) -> Any:
         """
         Get breakdown of SourceRank score for a given project.
 
@@ -108,7 +106,8 @@ class Search(object):
 
         return search_api("project_sourcerank", platforms, project)
 
-    def project_usage(self, platforms: str, project: str) -> Any:
+    @staticmethod
+    def project_usage(platforms: str, project: str) -> Any:
         """
         Get breakdown of usage for a given project.
 
@@ -121,7 +120,8 @@ class Search(object):
 
         return search_api("project_usage", platforms, project)
 
-    def project_search(self, **kwargs):
+    @staticmethod
+    def project_search(**kwargs):
         """
         Search for projects.
         Args - keywords only:
@@ -139,7 +139,8 @@ class Search(object):
         """
         return search_api("special_project_search", **kwargs)
 
-    def repository(self, host: str, owner: str, repo: str) -> Any:
+    @staticmethod
+    def repository(host: str, owner: str, repo: str) -> Any:
         """
         Return information about a repository and its versions.
 
@@ -153,7 +154,8 @@ class Search(object):
 
         return search_api("repository", host, owner, repo)
 
-    def repository_dependencies(self, host: str, owner: str, repo: str) -> Any:
+    @staticmethod
+    def repository_dependencies(host: str, owner: str, repo: str) -> Any:
         """
         Return information about a repository's dependencies.
 
@@ -167,7 +169,8 @@ class Search(object):
 
         return search_api("repository_dependencies", host, owner, repo)
 
-    def repository_projects(self, host: str, owner: str, repo: str) -> Any:
+    @staticmethod
+    def repository_projects(host: str, owner: str, repo: str) -> Any:
         """
         Get a list of projects referencing the given repository.
 
@@ -181,7 +184,8 @@ class Search(object):
 
         return search_api("repository_projects", host, owner, repo)
 
-    def user(self, host: str, user: str) -> Any:
+    @staticmethod
+    def user(host: str, user: str) -> Any:
         """
         Return information about a user.
 
@@ -193,7 +197,8 @@ class Search(object):
         """
         return search_api("user", host, user)
 
-    def user_repositories(self, host: str, user: str) -> Any:
+    @staticmethod
+    def user_repositories(host: str, user: str) -> Any:
         """
         Return information about a user's repos.
 
@@ -205,7 +210,8 @@ class Search(object):
         """
         return search_api("user_repositories", host, user)
 
-    def user_projects(self, host: str, user: str) -> Any:
+    @staticmethod
+    def user_projects(host: str, user: str) -> Any:
         """
         Return information about projects using a user's repos.
 
@@ -217,7 +223,8 @@ class Search(object):
         """
         return search_api("user_projects", host, user)
 
-    def user_projects_contributions(self, host: str, user: str) -> Any:
+    @staticmethod
+    def user_projects_contributions(host: str, user: str) -> Any:
         """
         Return information about projects a user has contributed to.
 
@@ -229,7 +236,8 @@ class Search(object):
         """
         return search_api("user_projects_contributions", host, user)
 
-    def user_repository_contributions(self, host: str, user: str) -> Any:
+    @staticmethod
+    def user_repository_contributions(host: str, user: str) -> Any:
         """
         Return information about repositories a user has contributed to.
 
@@ -241,7 +249,8 @@ class Search(object):
         """
         return search_api("user_repositories_contributions", host, user)
 
-    def user_dependencies(self, host, user):
+    @staticmethod
+    def user_dependencies(host, user):
         """
         Return a list of unique user's repositories' dependencies.