+++ tests/test_subscribe.py
@@ -2,6 +2,7 @@
 from time import sleep
 
 import pytest
+
 import pybraries
 
 
@@ -24,39 +25,41 @@ repo3 = "scikit-learn"  # repo name
 # won't pass locally for development if user search_key isn't subscribed to any packages
 @pytest.mark.xfail(raises=IndexError)
 def test_list_subscribed():
-    """for subscriber- returns a list item subscription with a project rank >= 0"""
+    """for subscriber, returns a list item subscription with a project rank >= 0"""
     sub = subs.list_subscribed()
     assert sub[0]["project"]["rank"] >= 0
 
 
 def test_check_subscribed():
-    """for subscriber - check if user is subscribed to a project with a rank >= 0"""
+    """for subscriber, check if user is subscribed to a project with a rank >= 0"""
     check_sub = subs.check_subscribed(mgr, repo3)
     assert type(check_sub) is bool
 
 
 def test_subscribe(pre_unsub):
-    """for subscribe key sent- use args to check if subscribed"""
+    """for subscribe key sent, use args to check if subscribed"""
     sub = subs.subscribe(mgr, repo2)
     assert type(sub) is str
 
 
 def test_subscribe_kwargs(pre_unsub):
-    """for api key sent- use kwargs to check if subscribed"""
+    """for api key sent, use kwargs to check if subscribed"""
     sub = subs.subscribe(manager=mgr, package=repo2)
     assert type(sub) is str
 
 
 @pytest.mark.skip()
 def test_update_subscribe():
-    """for api key sent- doesn't error"""
+    """for api key sent,  doesn't error"""
     update = subs.update_subscribe(mgr, repo2)
-    pass
+
+    if not update:
+        print("Something bad happened")
 
 
 @pytest.mark.skip()
 def test_update_subscribe_updates():
-    """for api key sent- change subscription for prerelease to false"""
+    """for api key sent, change subscription for prerelease to false"""
     update = subs.update_subscribe(mgr, repo2, False)
     assert update["include_prerelease"] is False
 
@@ -75,7 +78,7 @@ def test_unsubscribe_intercept(pre_unsub):
 
 # slow
 def test_unsubscribe_args(pre_sub):
-    """for api key sent- doesn't error if not already subscribed"""
+    """for api key sent, doesn't error if not already subscribed"""
     del_sub = subs.unsubscribe(mgr, repo2)
     assert del_sub == "Successfully Unsubscribed"
 
@@ -83,5 +86,9 @@ def test_unsubscribe_args(pre_sub):
 def test_unsubscribe_works():
     """unsubscribes and verifies"""
     del_sub = subs.unsubscribe(mgr, repo2)
+
+    if not del_sub:
+        print("Something went horribly wrong.")
+
     bsub = subs.check_subscribed(mgr, repo2)
     assert bsub is False